fixme

+++ main.py
@@ -108,7 +108,7 @@ class App:
         self.bk.quit()
         self.bk.destroy()
         self.exit_code = 0
-        if not sys.platform == 'win32':
+        if not sys.platform == 'win32' and hasattr(self.server, 'sock'):
             os.kill(os.getpid(), 9)  # FIXME
 
     def track_loop(self) -> None:
